* check if log file exists before delete
git-svn-id: https://shiva.ms.mff.cuni.cz/svn/zlomekfs/branches/zouhar@2976 b158dfce-25fb-0310-9a5b-f29f65ed45b5

+++ insecticide/snapshotPlugin.py
@@ -300,7 +300,7 @@ class SnapshotPlugin(Plugin):
             rootLogger.removeHandler(self.noseLogHandler)
             self.noseLogHandler.close()
             self.noseLogHandler = None
-        if self.noseLogFileName:
+        if self.noseLogFileName and os.path.isfile(self.noseLogFileName):
             os.unlink(self.noseLogFileName)
             for backLogNumber in range(self.noseLogBackupCount, 
                 self.noseLogBackupCount + 1):
